EC-831 - updates based on the code review.
